refactor(season): Make season named widget slice on client
- Made season named widget slice
- Removed SeasonList to season named widget slice
