+++ packages/client/src/widgets/season/ui/index.ts
@@ -1,0 +1,1 @@
+export { SeasonList } from './list';
